# Conflicts:
#	erp-server/erp-server-mrp/src/main/java/com/erp/server/mrp/es/service/OrderHistorySalesEsService.java
#	erp-server/erp-server-mrp/src/main/java/com/erp/server/mrp/es/service/impl/OrderHistorySalesEsServiceImpl.java
#	erp-server/erp-server-mrp/src/main/java/com/erp/server/mrp/service/impl/CalcSalesInfoDimServiceImpl.java

+++ erp-server/erp-server-wms/src/main/java/com/erp/server/wms/service/impl/ReportOrderDataServiceImpl.java
@@ -142,7 +142,7 @@ public class ReportOrderDataServiceImpl extends SuperServiceImpl<ReportOrderData
           throw new ServiceException("已有任务进行中，请勿重复提交请求");
         }
         //添加缓存
-        redisUtil.set(existKey,isAuto, RedisCacheConstants.LOCK_DURATION_MINUTES * 10);
+        redisUtil.set(existKey,isAuto, (long) RedisCacheConstants.LOCK_DURATION_MINUTES * 10);
         //生成缺货统计、销售看板
         try {
             generateAllReport(viewDTO);
